+++ Tests/Shared/StaticDelegateRegistry.generated.cs
@@ -57,7 +57,7 @@ namespace Burst.Compiler.IL.Tests
             Register(typeof(byte),  new Type[] {typeof(uint)}, typeof(Burst_Compiler_IL_Tests_Expressions_ConvOvfU1UnDelegate), CallBurst_Compiler_IL_Tests_Expressions_ConvOvfU1UnDelegate);
             Register(typeof(double),  new Type[] {typeof(Unity.Mathematics.double2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_AbsDouble2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_AbsDouble2Delegate);
             Register(typeof(double),  new Type[] {typeof(Unity.Mathematics.double3).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_AbsDouble3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_AbsDouble3Delegate);
-            Register(typeof(double),  new Type[] {typeof(Unity.Mathematics.double4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_AbsDouble4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_AbsDouble4Delegate);
+            Register(typeof(double),  new Type[] {typeof(Unity.Mathematics.double4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Cos4dDelegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Cos4dDelegate);
             Register(typeof(double),  new Type[] {typeof(Unity.Mathematics.int4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Doubles_Double4Int4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Doubles_Double4Int4Delegate);
             Register(typeof(double),  new Type[] {typeof(double), typeof(double)}, typeof(Burst_Compiler_IL_Tests_TestSystemMath_TestMaxDoubleDelegate), CallBurst_Compiler_IL_Tests_TestSystemMath_TestMaxDoubleDelegate);
             Register(typeof(double),  new Type[] {typeof(double)}, typeof(Burst_Compiler_IL_Tests_TestUnityMath_TestSignDoubleDelegate), CallBurst_Compiler_IL_Tests_TestUnityMath_TestSignDoubleDelegate);
@@ -71,22 +71,22 @@ namespace Burst.Compiler.IL.Tests
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.bool3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Select3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Select3Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.bool4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Select4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Select4Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(Unity.Mathematics.float2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Lerp2_2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp2_2Delegate);
-            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Lerp2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp2Delegate);
-            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(Unity.Mathematics.float2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Min2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Min2Delegate);
-            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Saturate2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Saturate2Delegate);
+            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Smoothstep2Delegate);
+            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float2).MakeByRefType(), typeof(Unity.Mathematics.float2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_ATan2_2Delegate);
+            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Cos2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Cos2Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Lerp3_3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp3_3Delegate);
-            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Lerp3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp3Delegate);
-            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Min3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Min3Delegate);
+            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Smoothstep3Delegate);
+            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float3).MakeByRefType(), typeof(Unity.Mathematics.float3).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_ATan2_3Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float3).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Indexers_Float3_get_IndexerByRefDelegate), CallBurst_Compiler_IL_Tests_Vectors_Indexers_Float3_get_IndexerByRefDelegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Lerp4_4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp4_4Delegate);
-            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Lerp4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp4Delegate);
-            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Min4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Min4Delegate);
+            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_Smoothstep4Delegate);
+            Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float4).MakeByRefType(), typeof(Unity.Mathematics.float4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_ATan2_4Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.float4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Deterministic_CosFloat4Delegate), CallBurst_Compiler_IL_Tests_Deterministic_CosFloat4Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.int2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_AsFloatInt2Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_AsFloatInt2Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.int3).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_AsFloatInt3Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_AsFloatInt3Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.int4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Constructors_Float4Int4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Constructors_Float4Int4Delegate);
             Register(typeof(float),  new Type[] {typeof(Unity.Mathematics.uint4).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_AsFloatUInt4Delegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_AsFloatUInt4Delegate);
-            Register(typeof(float),  new Type[] {typeof(float), typeof(float), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_LerpDelegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_LerpDelegate);
+            Register(typeof(float),  new Type[] {typeof(float), typeof(float), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Vectors_Maths_SmoothstepDelegate), CallBurst_Compiler_IL_Tests_Vectors_Maths_SmoothstepDelegate);
             Register(typeof(float),  new Type[] {typeof(float), typeof(float)}, typeof(Burst_Compiler_IL_Tests_Expressions_test_expr_add_two_argumentsDelegate), CallBurst_Compiler_IL_Tests_Expressions_test_expr_add_two_argumentsDelegate);
             Register(typeof(float),  new Type[] {typeof(float)}, typeof(Burst_Compiler_IL_Tests_Expressions_test_expr_add_multiplesDelegate), CallBurst_Compiler_IL_Tests_Expressions_test_expr_add_multiplesDelegate);
             Register(typeof(float),  new Type[] {typeof(int), typeof(int)}, typeof(Burst_Compiler_IL_Tests_ControlFlows_ForWhileNestedCallDelegate), CallBurst_Compiler_IL_Tests_ControlFlows_ForWhileNestedCallDelegate);
@@ -141,7 +141,7 @@ namespace Burst.Compiler.IL.Tests
             Register(typeof(int),  new Type[] {typeof(int), typeof(int)}, typeof(Burst_Compiler_IL_Tests_Expressions_BinaryNumericAddDelegate), CallBurst_Compiler_IL_Tests_Expressions_BinaryNumericAddDelegate);
             Register(typeof(int),  new Type[] {typeof(int).MakeByRefType(), typeof(int), typeof(int)}, typeof(Burst_Compiler_IL_Tests_TestAtomics_test_atomic_compare_and_exchange_intDelegate), CallBurst_Compiler_IL_Tests_TestAtomics_test_atomic_compare_and_exchange_intDelegate);
             Register(typeof(int),  new Type[] {typeof(int).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_ControlFlows_WhileDynamicDelegate), CallBurst_Compiler_IL_Tests_ControlFlows_WhileDynamicDelegate);
-            Register(typeof(int),  new Type[] {typeof(int)}, typeof(Burst_Compiler_IL_Tests_Expressions_AddOverflowIntDelegate), CallBurst_Compiler_IL_Tests_Expressions_AddOverflowIntDelegate);
+            Register(typeof(int),  new Type[] {typeof(int)}, typeof(Burst_Compiler_IL_Tests_Expressions_BltDelegate), CallBurst_Compiler_IL_Tests_Expressions_BltDelegate);
             Register(typeof(int),  new Type[] {typeof(int*), typeof(int)}, typeof(Burst_Compiler_IL_Tests_Pointers_UnsafeCompareDelegate), CallBurst_Compiler_IL_Tests_Pointers_UnsafeCompareDelegate);
             Register(typeof(int),  new Type[] {typeof(long)}, typeof(Burst_Compiler_IL_Tests_Expressions_ConvOvfI4Delegate), CallBurst_Compiler_IL_Tests_Expressions_ConvOvfI4Delegate);
             Register(typeof(int),  new Type[] {typeof(short)}, typeof(Burst_Compiler_IL_Tests_Expressions_ConvertShortToIntDelegate), CallBurst_Compiler_IL_Tests_Expressions_ConvertShortToIntDelegate);
@@ -171,7 +171,7 @@ namespace Burst.Compiler.IL.Tests
             Register(typeof(uint),  new Type[] {typeof(int)}, typeof(Burst_Compiler_IL_Tests_Expressions_ConvOvfU4Delegate), CallBurst_Compiler_IL_Tests_Expressions_ConvOvfU4Delegate);
             Register(typeof(uint),  new Type[] {typeof(sbyte)}, typeof(Burst_Compiler_IL_Tests_Expressions_ConvertSbyteToUIntDelegate), CallBurst_Compiler_IL_Tests_Expressions_ConvertSbyteToUIntDelegate);
             Register(typeof(uint),  new Type[] {typeof(uint), typeof(int)}, typeof(Burst_Compiler_IL_Tests_Expressions_BinaryNumericShiftLeftUInt32Delegate), CallBurst_Compiler_IL_Tests_Expressions_BinaryNumericShiftLeftUInt32Delegate);
-            Register(typeof(uint),  new Type[] {typeof(uint)}, typeof(Burst_Compiler_IL_Tests_Expressions_AddOvfUnDelegate), CallBurst_Compiler_IL_Tests_Expressions_AddOvfUnDelegate);
+            Register(typeof(uint),  new Type[] {typeof(uint)}, typeof(Burst_Compiler_IL_Tests_Expressions_BgtUnDelegate), CallBurst_Compiler_IL_Tests_Expressions_BgtUnDelegate);
             Register(typeof(uint),  new Type[] {typeof(ulong)}, typeof(Burst_Compiler_IL_Tests_Expressions_ConvOvfU4UnDelegate), CallBurst_Compiler_IL_Tests_Expressions_ConvOvfU4UnDelegate);
             Register(typeof(ulong),  Array.Empty<Type>(), typeof(Burst_Compiler_IL_Tests_Expressions_ConvertIntToLongConstDelegate), CallBurst_Compiler_IL_Tests_Expressions_ConvertIntToLongConstDelegate);
             Register(typeof(ulong),  new Type[] {typeof(double)}, typeof(Burst_Compiler_IL_Tests_Expressions_ConvOvfU8Delegate), CallBurst_Compiler_IL_Tests_Expressions_ConvOvfU8Delegate);
@@ -226,6 +226,7 @@ namespace Burst.Compiler.IL.Tests
             Register(typeof(int),  new Type[] {typeof(Burst.Compiler.IL.Tests.NativeContainers.NativeSlices.NativeSliceIntCopy).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_NativeContainers_NativeSlices_test_expr_native_slice_int_copy_sliceDelegate), CallBurst_Compiler_IL_Tests_NativeContainers_NativeSlices_test_expr_native_slice_int_copy_sliceDelegate);
             Register(typeof(int),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestCullingJob.CullingJob).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestCullingJob_test_culling_job_simpleDelegate), CallBurst_Compiler_IL_Tests_TestCullingJob_test_culling_job_simpleDelegate);
             Register(typeof(int),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestFunctionPointers.FPBigStructFunc), typeof(Burst.Compiler.IL.Tests.ABI.BigStruct)}, typeof(Burst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerWithBigStructDelegate), CallBurst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerWithBigStructDelegate);
+            Register(typeof(int),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestFunctionPointers.FPInIntFunc), typeof(int), typeof(int)}, typeof(Burst_Compiler_IL_Tests_TestFunctionPointers_TestAddTwoNumbersByInDelegate), CallBurst_Compiler_IL_Tests_TestFunctionPointers_TestAddTwoNumbersByInDelegate);
             Register(typeof(int),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestFunctionPointers.FPIntFunc), typeof(Burst.Compiler.IL.Tests.TestFunctionPointers.FPIntFunc), typeof(Burst.Compiler.IL.Tests.NativeContainers.NativeArrays.NativeArrayFloat).MakeByRefType(), typeof(int)}, typeof(Burst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerDynamicInvokeDelegate), CallBurst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerDynamicInvokeDelegate);
             Register(typeof(int),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestFunctionPointers.FPIntFunc), typeof(Burst.Compiler.IL.Tests.TestFunctionPointers.FPIntFunc), typeof(int)}, typeof(Burst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerInvokeDelegate), CallBurst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerInvokeDelegate);
             Register(typeof(int),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestGenerics.TestJob).MakeByRefType(), typeof(int)}, typeof(Burst_Compiler_IL_Tests_TestGenerics_ExecuteAudioBugDelegate), CallBurst_Compiler_IL_Tests_TestGenerics_ExecuteAudioBugDelegate);
@@ -250,6 +251,7 @@ namespace Burst.Compiler.IL.Tests
             Register(typeof(void),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestAlias.SimpleSOA1Job).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestAlias_TestSimpleSOA1JobDelegate), CallBurst_Compiler_IL_Tests_TestAlias_TestSimpleSOA1JobDelegate);
             Register(typeof(void),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestGenerics.StructWithGenerics).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestGenerics_TestStructWithGenericsDelegate), CallBurst_Compiler_IL_Tests_TestGenerics_TestStructWithGenericsDelegate);
             Register(typeof(void),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestGenerics.StructWithGenerics2).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestGenerics_TestStructWithGenerics2Delegate), CallBurst_Compiler_IL_Tests_TestGenerics_TestStructWithGenerics2Delegate);
+            Register(typeof(void),  new Type[] {typeof(Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000).MakeByRefType(), typeof(Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000).MakeByRefType(), typeof(Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestGenerics_TestSymbolLengthDelegate), CallBurst_Compiler_IL_Tests_TestGenerics_TestSymbolLengthDelegate);
             Register(typeof(void),  new Type[] {typeof(UnityBenchShared.DivisionByScalar).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestPerfNotes_test_division_by_scalarDelegate), CallBurst_Compiler_IL_Tests_TestPerfNotes_test_division_by_scalarDelegate);
             Register(typeof(void),  new Type[] {typeof(UnityBenchShared.IntToFloatPenalty).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestPerfNotes_test_int_to_float_penaltyDelegate), CallBurst_Compiler_IL_Tests_TestPerfNotes_test_int_to_float_penaltyDelegate);
             Register(typeof(void),  new Type[] {typeof(UnityBenchShared.PartialWrite).MakeByRefType()}, typeof(Burst_Compiler_IL_Tests_TestPerfNotes_test_partial_writeDelegate), CallBurst_Compiler_IL_Tests_TestPerfNotes_test_partial_writeDelegate);
@@ -727,10 +729,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate double Burst_Compiler_IL_Tests_Vectors_Maths_AbsDouble4Delegate(ref Unity.Mathematics.double4 arg0);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_AbsDouble4Delegate(object delegateObj, object[] p)
+        private delegate double Burst_Compiler_IL_Tests_Vectors_Maths_Cos4dDelegate(ref Unity.Mathematics.double4 arg0);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Cos4dDelegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_AbsDouble4Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Cos4dDelegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.double4)p[0];
             result = d(ref arg0);
@@ -876,10 +878,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Lerp2Delegate(ref Unity.Mathematics.float2 arg0, ref Unity.Mathematics.float2 arg1, float arg2);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp2Delegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep2Delegate(ref Unity.Mathematics.float2 arg0, ref Unity.Mathematics.float2 arg1, float arg2);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Smoothstep2Delegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Lerp2Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep2Delegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.float2)p[0];
             var arg1 = (Unity.Mathematics.float2)p[1];
@@ -888,10 +890,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Min2Delegate(ref Unity.Mathematics.float2 arg0, ref Unity.Mathematics.float2 arg1);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Min2Delegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_2Delegate(ref Unity.Mathematics.float2 arg0, ref Unity.Mathematics.float2 arg1);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_ATan2_2Delegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Min2Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_2Delegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.float2)p[0];
             var arg1 = (Unity.Mathematics.float2)p[1];
@@ -899,10 +901,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Saturate2Delegate(ref Unity.Mathematics.float2 arg0);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Saturate2Delegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Cos2Delegate(ref Unity.Mathematics.float2 arg0);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Cos2Delegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Saturate2Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Cos2Delegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.float2)p[0];
             result = d(ref arg0);
@@ -921,10 +923,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Lerp3Delegate(ref Unity.Mathematics.float3 arg0, ref Unity.Mathematics.float3 arg1, float arg2);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp3Delegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep3Delegate(ref Unity.Mathematics.float3 arg0, ref Unity.Mathematics.float3 arg1, float arg2);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Smoothstep3Delegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Lerp3Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep3Delegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.float3)p[0];
             var arg1 = (Unity.Mathematics.float3)p[1];
@@ -933,10 +935,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Min3Delegate(ref Unity.Mathematics.float3 arg0, ref Unity.Mathematics.float3 arg1);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Min3Delegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_3Delegate(ref Unity.Mathematics.float3 arg0, ref Unity.Mathematics.float3 arg1);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_ATan2_3Delegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Min3Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_3Delegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.float3)p[0];
             var arg1 = (Unity.Mathematics.float3)p[1];
@@ -966,10 +968,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Lerp4Delegate(ref Unity.Mathematics.float4 arg0, ref Unity.Mathematics.float4 arg1, float arg2);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Lerp4Delegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep4Delegate(ref Unity.Mathematics.float4 arg0, ref Unity.Mathematics.float4 arg1, float arg2);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Smoothstep4Delegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Lerp4Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Smoothstep4Delegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.float4)p[0];
             var arg1 = (Unity.Mathematics.float4)p[1];
@@ -978,10 +980,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_Min4Delegate(ref Unity.Mathematics.float4 arg0, ref Unity.Mathematics.float4 arg1);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_Min4Delegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_4Delegate(ref Unity.Mathematics.float4 arg0, ref Unity.Mathematics.float4 arg1);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_ATan2_4Delegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_Min4Delegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_ATan2_4Delegate)delegateObj;
             object result = null;
             var arg0 = (Unity.Mathematics.float4)p[0];
             var arg1 = (Unity.Mathematics.float4)p[1];
@@ -1039,10 +1041,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_LerpDelegate(float arg0, float arg1, float arg2);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_LerpDelegate(object delegateObj, object[] p)
+        private delegate float Burst_Compiler_IL_Tests_Vectors_Maths_SmoothstepDelegate(float arg0, float arg1, float arg2);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Vectors_Maths_SmoothstepDelegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_LerpDelegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Vectors_Maths_SmoothstepDelegate)delegateObj;
             object result = null;
             var arg0 = (float)p[0];
             var arg1 = (float)p[1];
@@ -1625,10 +1627,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate int Burst_Compiler_IL_Tests_Expressions_AddOverflowIntDelegate(int arg0);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Expressions_AddOverflowIntDelegate(object delegateObj, object[] p)
+        private delegate int Burst_Compiler_IL_Tests_Expressions_BltDelegate(int arg0);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Expressions_BltDelegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Expressions_AddOverflowIntDelegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Expressions_BltDelegate)delegateObj;
             object result = null;
             var arg0 = (int)p[0];
             result = d(arg0);
@@ -1930,10 +1932,10 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-        private delegate uint Burst_Compiler_IL_Tests_Expressions_AddOvfUnDelegate(uint arg0);
-        private unsafe static object CallBurst_Compiler_IL_Tests_Expressions_AddOvfUnDelegate(object delegateObj, object[] p)
+        private delegate uint Burst_Compiler_IL_Tests_Expressions_BgtUnDelegate(uint arg0);
+        private unsafe static object CallBurst_Compiler_IL_Tests_Expressions_BgtUnDelegate(object delegateObj, object[] p)
         {
-            var d = (Burst_Compiler_IL_Tests_Expressions_AddOvfUnDelegate)delegateObj;
+            var d = (Burst_Compiler_IL_Tests_Expressions_BgtUnDelegate)delegateObj;
             object result = null;
             var arg0 = (uint)p[0];
             result = d(arg0);
@@ -2462,6 +2464,18 @@ namespace Burst.Compiler.IL.Tests
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
+        private delegate int Burst_Compiler_IL_Tests_TestFunctionPointers_TestAddTwoNumbersByInDelegate(Burst.Compiler.IL.Tests.TestFunctionPointers.FPInIntFunc arg0, int arg1, int arg2);
+        private unsafe static object CallBurst_Compiler_IL_Tests_TestFunctionPointers_TestAddTwoNumbersByInDelegate(object delegateObj, object[] p)
+        {
+            var d = (Burst_Compiler_IL_Tests_TestFunctionPointers_TestAddTwoNumbersByInDelegate)delegateObj;
+            object result = null;
+            var arg0 = (Burst.Compiler.IL.Tests.TestFunctionPointers.FPInIntFunc)p[0];
+            var arg1 = (int)p[1];
+            var arg2 = (int)p[2];
+            result = d(arg0, arg1, arg2);
+            return result;
+        }
+        [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
         private delegate int Burst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerDynamicInvokeDelegate(Burst.Compiler.IL.Tests.TestFunctionPointers.FPIntFunc arg0, Burst.Compiler.IL.Tests.TestFunctionPointers.FPIntFunc arg1, ref Burst.Compiler.IL.Tests.NativeContainers.NativeArrays.NativeArrayFloat arg2, int arg3);
         private unsafe static object CallBurst_Compiler_IL_Tests_TestFunctionPointers_TestFunctionPointerDynamicInvokeDelegate(object delegateObj, object[] p)
         {
@@ -2712,6 +2726,18 @@ namespace Burst.Compiler.IL.Tests
             object result = null;
             var arg0 = (Burst.Compiler.IL.Tests.TestGenerics.StructWithGenerics2)p[0];
             d(ref arg0);
+            return result;
+        }
+        [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
+        private delegate void Burst_Compiler_IL_Tests_TestGenerics_TestSymbolLengthDelegate(ref Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 arg0, ref Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 arg1, ref Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 arg2);
+        private unsafe static object CallBurst_Compiler_IL_Tests_TestGenerics_TestSymbolLengthDelegate(object delegateObj, object[] p)
+        {
+            var d = (Burst_Compiler_IL_Tests_TestGenerics_TestSymbolLengthDelegate)delegateObj;
+            object result = null;
+            var arg0 = (Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000)p[0];
+            var arg1 = (Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000)p[1];
+            var arg2 = (Burst.Compiler.IL.Tests.TestGenerics.StructWithGenericsAndALongName00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000)p[2];
+            d(ref arg0, ref arg1, ref arg2);
             return result;
         }
         [UnmanagedFunctionPointer(CallingConvention.Cdecl)]